+++ Citadel/src/utopia/citadel/database/access/many/description/DbDescriptionRoles.scala
@@ -1,8 +1,17 @@
 package utopia.citadel.database.access.many.description
 
+import utopia.citadel.util.CitadelContext
+import utopia.flow.datastructure.mutable.{ExpiringLazy, RefreshingLazy}
+import utopia.flow.datastructure.template.LazyLike
+import utopia.flow.time.TimeExtensions._
+import utopia.flow.util.CollectionExtensions._
 import utopia.metropolis.model.combined.description.DescribedDescriptionRole
 import utopia.metropolis.model.stored.description.DescriptionRole
+import utopia.vault.database.Connection
 import utopia.vault.nosql.view.UnconditionalView
+import utopia.vault.util.ErrorHandling
+
+import scala.concurrent.duration.Duration
 
 /**
   * The root access point when targeting multiple DescriptionRoles at a time
@@ -11,6 +20,72 @@ import utopia.vault.nosql.view.UnconditionalView
   */
 object DbDescriptionRoles extends ManyDescriptionRolesAccess with UnconditionalView
 {
+	// ATTRIBUTES   ----------------
+	
+	// Caches all description roles, since they are needed relatively often and change rarely
+	private lazy val cache: Option[LazyLike[Vector[DescriptionRole]]] =
+	{
+		val cacheDuration = CitadelContext.descriptionRoleCacheDuration
+		if (cacheDuration <= Duration.Zero)
+			None
+		else
+		{
+			import CitadelContext.executionContext
+			import CitadelContext.connectionPool
+			
+			def _readValues = connectionPool.tryWith { implicit c => pull }
+				.getOrMap { error =>
+					ErrorHandling.defaultPrinciple.handle(error)
+					Vector()
+				}
+			
+			if (cacheDuration > 1.minutes)
+				Some(RefreshingLazy.after(cacheDuration)(_readValues))
+			else
+				Some(ExpiringLazy.after(cacheDuration)(_readValues))
+		}
+	}
+	
+	
+	// IMPLEMENTED  ------------------
+	
+	override def ids(implicit connection: Connection) = cache match
+	{
+		case Some(c) => c.value.map { _.id }
+		case None => super.ids
+	}
+	override def size(implicit connection: Connection) = cache match
+	{
+		case Some(c) => c.value.size
+		case None => super.size
+	}
+	override def all(implicit connection: Connection) = cache match
+	{
+		case Some(c) => c.value
+		case None => super.all
+	}
+	override def foreach[U](f: DescriptionRole => U)(implicit connection: Connection) = cache match
+	{
+		case Some(c) => c.value.foreach(f)
+		case None => super.foreach(f)
+	}
+	override def pull(implicit connection: Connection) = cache match
+	{
+		case Some(c) => c.value
+		case None => super.pull
+	}
+	override def nonEmpty(implicit connection: Connection) = cache match
+	{
+		case Some(c) => c.value.nonEmpty
+		case None => super.nonEmpty
+	}
+	override def isEmpty(implicit connection: Connection) = cache match
+	{
+		case Some(c) => c.value.isEmpty
+		case None => super.isEmpty
+	}
+	
+	
 	// OTHER	--------------------
 	
 	/**