+++ Citadel/src/utopia/citadel/util/CitadelContext.scala
@@ -1,9 +1,11 @@
 package utopia.citadel.util
 
 import utopia.flow.generic.{DataType, EnvironmentNotSetupException}
+import utopia.flow.time.TimeExtensions._
 import utopia.vault.database.ConnectionPool
 
 import scala.concurrent.ExecutionContext
+import scala.concurrent.duration.Duration
 
 /**
   * This context object holds certain global values required by many features in this project. It is recommended to
@@ -26,20 +28,26 @@ object CitadelContext
 	  */
 	@throws[EnvironmentNotSetupException](".setup(...) hasn't been called yet")
 	implicit def executionContext: ExecutionContext = get.exc
-	
 	/**
 	  * @return Database connection pool to use in this project (implicit)
 	  * @throws EnvironmentNotSetupException If setup(...) hasn't been called yet
 	  */
 	@throws[EnvironmentNotSetupException](".setup(...) hasn't been called yet")
 	implicit def connectionPool: ConnectionPool = get.connectionPool
-	
 	/**
 	  * @return Name of the database that contains the project specific tables
 	  * @throws EnvironmentNotSetupException If .setup(...) hasn't been called yet
 	  */
 	@throws[EnvironmentNotSetupException](".setup(...) hasn't been called yet")
 	def databaseName = get.databaseName
+	/**
+	  * @return Duration how long description roles may be cached outside of the database
+	  */
+	def descriptionRoleCacheDuration = data match
+	{
+		case Some(data) => data.descriptionRoleCacheDuration
+		case None => Duration.Zero
+	}
 	
 	@throws[EnvironmentNotSetupException]("If .setup(...) hasn't been called yet")
 	private def get = data match
@@ -56,15 +64,21 @@ object CitadelContext
 	  * @param executionContext Execution context used
 	  * @param connectionPool Database connection pool used
 	  * @param databaseName Name of the primary database where data is stored
+	  * @param descriptionRoleCacheDuration Duration how long description roles may be cached so that they don't
+	  *                                     need to be read from the database. Use a lower value if you expect
+	  *                                     changes in description role database data. If there are no changes ever,
+	  *                                     infinite duration may be appropriate. (Default = 1 hours)
 	  */
-	def setup(executionContext: ExecutionContext, connectionPool: ConnectionPool, databaseName: String) =
+	def setup(executionContext: ExecutionContext, connectionPool: ConnectionPool, databaseName: String,
+	          descriptionRoleCacheDuration: Duration = 1.hours) =
 	{
 		DataType.setup()
-		data = Some(Data(executionContext, connectionPool, databaseName))
+		data = Some(Data(executionContext, connectionPool, databaseName, descriptionRoleCacheDuration))
 	}
 	
 	
 	// NESTED	--------------------------------------
 	
-	private case class Data(exc: ExecutionContext, connectionPool: ConnectionPool, databaseName: String)
+	private case class Data(exc: ExecutionContext, connectionPool: ConnectionPool, databaseName: String,
+	                        descriptionRoleCacheDuration: Duration)
 }