+++ Flow/src/utopia/flow/datastructure/mutable/ExpiringLazy.scala
@@ -1,6 +1,7 @@
 package utopia.flow.datastructure.mutable
 
 import utopia.flow.async.ResettableVolatileLazy
+import utopia.flow.datastructure.immutable.Lazy
 import utopia.flow.time.{Now, WaitUtils}
 import utopia.flow.time.TimeExtensions._
 
@@ -21,6 +22,22 @@ object ExpiringLazy
 	  */
 	def after[A](expirationThreshold: FiniteDuration)(make: => A)(implicit exc: ExecutionContext) =
 		new ExpiringLazy[A](make)(_ => expirationThreshold)
+	/**
+	  * Creates a new lazy container that automatically resets its contents after
+	  * a specified duration has passed from value generation
+	  * @param expirationThreshold How long values are cached within this lazy -
+	  *                            If infinite, creates a standard lazy container
+	  * @param make Function for generating a new value when one is requested
+	  * @param exc Implicit execution context (used for scheduling the reset, if applicable)
+	  * @tparam A Type of cached value
+	  * @return A new lazy container
+	  */
+	def after[A](expirationThreshold: Duration)(make: => A)(implicit exc: ExecutionContext) =
+		expirationThreshold.finite match
+		{
+			case Some(duration) => new ExpiringLazy[A](make)(_ => duration)
+			case None => Lazy(make)
+		}
 	
 	/**
 	  * Creates a new lazy container that automatically resets its contents after
@@ -36,7 +53,7 @@ object ExpiringLazy
 }
 
 /**
-  * A lazy container that automatically resets its contents after a while
+  * A lazy container that automatically resets its contents after a while. Resetting is performed asynchronously.
   * @author Mikko Hilpinen
   * @since 16.5.2021, v1.10
   */